requirements.txt
Django==5.0.6
django-phonenumber-field[phonenumbers]==8.0.0
Babel==2.14.0
django-formset==1.4.4
django-countries==7.6.1
uvicorn-worker==0.2.0
uvicorn==0.30.1
gunicorn==22.0.0
whitenoise==6.7.0
psycopg2-binary==2.9.9
dj-database-url==2.2.0
django-admin-honeypot-updated-2021==1.2.0
django-otp==1.5.4
pytz==2024.2
beautifulsoup4==4.12.3
requests==2.32.3